+++ src/pages/logspage/LogsExplorer/LogsExplorer.tsx
